% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrete_pvalue}
\alias{discrete_pvalue}
\title{Convert continuous p-values into a factor based on the given split.}
\usage{
discrete_pvalue(p.value, split_by = c(0.001, 0.01, 0.05))
}
\arguments{
\item{p.value}{Vector of p-values.}

\item{split_by}{Where to split the p-values up by. Default is the <0.001,
<0.01, <0.05, and all others are >0.05.}
}
\value{
A factor vector of p-value splits.
}
\description{
Convert continuous p-values into a factor based on the given split.
}
\examples{

pvals <- c(0.001, 0.0123, 0.123, 0.009, 0.2145, 0.987, 0.05, 0.049, 0.051)
discrete_pvalue(pvals)
discrete_pvalue(pvals, split_by = c(0.05))
discrete_pvalue(pvals, split_by = c(0.01))
discrete_pvalue(pvals, split_by = c(0.01, 0.05))
discrete_pvalue(pvals, split_by = c(0.05, 0.001, 0.01))
}
