% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_corr_circle.R
\name{geom_corr_circle}
\alias{geom_corr_circle}
\title{Correlation circle plot for dimensionality reduction techniques.}
\usage{
geom_corr_circle(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  outer.linetype = "solid", outer.linecolour = "black",
  outer.linesize = 0.5, inner.linetype = "dotted",
  inner.linecolour = "black", inner.linesize = 0.5,
  center.linetype = "solid", center.linecolour = "grey50",
  center.linesize = 0.3, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{outer.linetype, outer.linecolour, outer.linesize}{The line type,
colour, and size for the outer circle line.}

\item{inner.linetype, inner.linecolour, inner.linesize}{The line type,
colour, and size for the inner circle line.}

\item{center.linetype, center.linecolour, center.linesize}{The line type,
colour, and size for the two center lines.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\value{
Adds a ggplot2 geom layer.
}
\description{
This creates a ggplot2 layer that plots the correlation values of the
scores for components against the real values, as obtained from
dimensionality reduction methods. These methods include principal components
analysis and partial least squares.
}
\examples{

library(ggplot2)
library(broom)
library(pls)
data(yarn)

# Set up data to plot.
fit <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
fit <- cor(model.matrix(fit), scores(fit)[, 1:2, drop = FALSE])
fit <- as.data.frame(fit)
fit$Variables <- rownames(fit)
rownames(fit) <- NULL
colnames(fit)[1:2] <- c("Comp1", "Comp2")
fit$Groups <- as.factor(sample.int(2, nrow(fit), replace = TRUE))

p <- ggplot(fit, aes(x = Comp1, y = Comp2))
p + geom_corr_circle()
p + geom_corr_circle(outer.linetype = "dotted")
p + geom_corr_circle(inner.linecolour = "blue")
p + geom_corr_circle(center.linesize = 1)
p + geom_corr_circle(center.linecolour = "grey50", size = 3)

# If you want to remove the circle or center lines, use 0.
p + geom_corr_circle(center.linetype = 0)
p + geom_corr_circle(outer.linetype = 0)

# With grouping
p + geom_corr_circle(aes(colour = Groups))
p + geom_corr_circle(aes(colour = Groups), size = 3) +
scale_colour_brewer()
p + geom_corr_circle(aes(alpha = Groups))
p + geom_corr_circle(aes(size = Groups))
p + geom_corr_circle(aes(shape = Groups), size = 2)

# With facets
p + geom_corr_circle() +
facet_grid(~ Groups)

}
