% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_estci.R, R/ggepi-package.R
\docType{data}
\name{GeomEstci}
\alias{GeomEstci}
\alias{ggepi-geoms}
\title{Base ggproto classes for ggepi}
\description{
Base ggproto classes for ggepi
}
\seealso{
\code{\link[ggplot2:ggproto]{ggplot2::ggproto()}}
}
\keyword{datasets}
\keyword{internal}
