% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_estci.R
\name{view_estci}
\alias{view_estci}
\title{Standard plot for estimates and confidence interval.}
\usage{
view_estci(data, xterms = "term", split_by = NA, ylab = "Predictor terms",
  xlab = "Estimate and CI", legend.title = "P-value", center.line = 0)
}
\arguments{
\item{data}{Tidied dataframe from a linear regression, as from the function
\link[broom]{tidy} from the broom package.}

\item{xterms}{Column that contains the predictor variables (ie. "term").}

\item{split_by}{Column to split the results up by (e.g. different models or
different y terms).}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label.}

\item{legend.title}{Legend title.}

\item{center.line}{Location for the center line.}
}
\value{
Plot of estimate and confidence interval.
}
\description{
Typical figure for examining estimates and confidence intervals from
regression type analyses. Has standard features for the plot, but if you want
more customization, see the \code{\link{geom_estci}} function.
}
\examples{

library(broom)
fit <- lm(Fertility ~ 0 + Catholic + Agriculture + Examination + Education + Infant.Mortality, data = swiss)
fit <- tidy(fit, conf.int = TRUE)
fit <- transform(fit, model = "non-log")

view_estci(fit, xterms = "term")
view_estci(fit, xterms = "term", center.line = 1)

fit_log <- lm(log(Fertility) ~ 0 + Catholic + Agriculture + Examination + Education + Infant.Mortality, data = swiss)
fit_log <- tidy(fit_log, conf.int = TRUE)
fit_log <- transform(fit_log, model = "log")
two_fits <- rbind(fit, fit_log)

view_estci(two_fits, xterms = "term", split_by = "model", center.line = 1)
}
\seealso{
geom_estci
}
